backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
httpx>=0.27.0
pydantic>=1.10.0
python-dotenv>=1.0.0
psycopg2-binary>=2.9.0
SQLAlchemy>=2.0.0
alpaca-py>=0.26.0
